Initial Import of CVS Stable branch.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/1.7@118 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServerScripts/playercommands/cmdhelp.cs
@@ -32,7 +32,7 @@ namespace DOL.GS.Scripts
 	{
 		public int OnCommand(GameClient client, string[] args)
 		{
-			uint plvl = (uint)client.Account.PrivLevel;
+			uint plvl = client.Account.PrivLevel;
 			bool iscmd = true;
 
 			if (args.Length > 1)
